requirements.txt
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
playwright>=1.40.0
jinja2>=3.1.2
python-dateutil>=2.8.2
requests>=2.31.0
